+++ functions/src/index.ts
@@ -11,3 +11,4 @@ setGlobalOptions({maxInstances: 10});
 export {onUserCreate} from "./handlers/authHandlers";
 export {saveGitHubToken} from "./handlers/serviceTokenHandlers";
 export {saveCard, getSavedCards} from "./handlers/savedCardHandlers";
+export {updateProfile} from "./handlers/profileHandlers";